feat: introduce ext that allows init resource once
